+++ Multiplicacion/src/main/java/edu/icesi/interfaces/IMatrixOperations.java
@@ -8,5 +8,5 @@ import org.osoa.sca.annotations.Service;
 @Service
 public interface IMatrixOperations{
 	public double[][] matrixMultiplication(double[][] m1,double[][] m2);
-	public List<long[][]> rotar(long[][] points, double angle);
+	public List<int[][]> rotar(int[][] points, double angle);
 }
